refactor: move icebridge data inputs to a separate module in io
