+++ scripts/MPI_median_elevation_filter.py
@@ -51,6 +51,7 @@ REFERENCE:
 
 UPDATE HISTORY:
     Updated 05/2023: using pathlib to define and operate on paths
+        move icebridge data inputs to a separate module in io
     Updated 12/2022: single implicit import of grounding zone tools
     Updated 07/2022: place some imports within try/except statements
     Updated 06/2022: updated ATM1b read functions for distributed version
@@ -107,11 +108,6 @@ import numpy as np
 import grounding_zones as gz
 
 # attempt imports
-try:
-    import ATM1b_QFIT.read_ATM1b_QFIT_binary
-except (ImportError, ModuleNotFoundError) as exc:
-    warnings.filterwarnings("module")
-    warnings.warn("ATM1b_QFIT not available", ImportWarning)
 try:
     import h5py
 except (ImportError, ModuleNotFoundError) as exc:
@@ -175,26 +171,6 @@ def set_hemisphere(REGION):
     projection_flag = {'AN':'S', 'GR':'N'}
     return projection_flag[REGION]
 
-# PURPOSE: reading the number of file lines removing commented lines
-def file_length(input_file,input_subsetter,HDF5=False,QFIT=False):
-    # subset the data to indices if specified
-    if input_subsetter:
-        file_lines = len(input_subsetter)
-    elif HDF5:
-        # read the size of an input variable within a HDF5 file
-        with h5py.File(input_file,'r') as fileID:
-            file_lines, = fileID[HDF5].shape
-    elif QFIT:
-        # read the size of a QFIT binary file
-        file_lines = ATM1b_QFIT.ATM1b_QFIT_shape(input_file)
-    else:
-        # read the input file, split at lines and remove all commented lines
-        with open(input_file, mode='r', encoding='utf8') as f:
-            i = [i for i in f.readlines() if re.match(r'^(?!\#|\n)',i)]
-        file_lines = len(i)
-    # return the number of lines
-    return file_lines
-
 # PURPOSE: read input HDF5 data from MPI_triangulate_elevation.py
 def read_HDF5_triangle_data(input_file, input_subsetter):
     # Open the HDF5 file for reading
@@ -219,276 +195,6 @@ def read_HDF5_triangle_data(input_file, input_subsetter):
     fileID.close()
     # return data and attributes
     return HDF5_data,HDF5_attributes
-
-# PURPOSE: read the ATM Level-1b data file for variables of interest
-def read_ATM_qfit_file(input_file, s):
-    # regular expression pattern for extracting parameters
-    mission_flag = r'(BLATM1B|ILATM1B|ILNSA1B)'
-    regex_pattern = rf'{mission_flag}_(\d+)_(\d+)(.*?).(qi|TXT|h5)'
-    # extract mission and other parameters from filename
-    MISSION,YYMMDD,HHMMSS,AUX,SFX = re.findall(regex_pattern,input_file).pop()
-    # early date strings omitted century and millenia (e.g. 93 for 1993)
-    if (len(YYMMDD) == 6):
-        yr2d,month,day = np.array([YYMMDD[:2],YYMMDD[2:4],YYMMDD[4:]],dtype='i')
-        year = (yr2d + 1900.0) if (yr2d >= 90) else (yr2d + 2000.0)
-    elif (len(YYMMDD) == 8):
-        year,month,day = np.array([YYMMDD[:4],YYMMDD[4:6],YYMMDD[6:]],dtype='i')
-    # output python dictionary with variables
-    ATM_L1b_input = {}
-    # Version 1 of ATM QFIT files (ascii)
-    # output text file from qi2txt with proper filename format
-    # do not use the shortened output format from qi2txt
-    if (SFX == 'TXT'):
-        # compile regular expression operator for reading lines
-        regex_pattern = r'[-+]?(?:(?:\d*\.\d+)|(?:\d+\.?))(?:[Ee][+-]?\d+)?'
-        rx = re.compile(regex_pattern, re.VERBOSE)
-        # read the input file, split at lines and remove all commented lines
-        f = open(input_file,'r')
-        file_contents=[i for i in f.readlines() if re.match(r'^(?!\#|\n)',i)]
-        f.close()
-        # number of lines of data within file
-        file_lines = file_length(input_file,s)
-        # subset the data to indices if specified
-        file_contents = [file_contents[i] for i in s] if s else file_contents
-        # create output variables with length equal to the number of lines
-        ATM_L1b_input['lat'] = np.zeros((file_lines),dtype=np.float64)
-        ATM_L1b_input['lon'] = np.zeros((file_lines),dtype=np.float64)
-        ATM_L1b_input['data'] = np.zeros((file_lines),dtype=np.float64)
-        hour = np.zeros((file_lines),dtype=np.float64)
-        minute = np.zeros((file_lines),dtype=np.float64)
-        second = np.zeros((file_lines),dtype=np.float64)
-        # for each line within the file
-        for i,line in enumerate(file_contents):
-            # find numerical instances within the line
-            line_contents = rx.findall(line)
-            ATM_L1b_input['lat'][i] = np.float64(line_contents[1])
-            ATM_L1b_input['lon'][i] = np.float64(line_contents[2])
-            ATM_L1b_input['data'][i] = np.float64(line_contents[3])
-            hour[i] = np.float64(line_contents[-1][:2])
-            minute[i] = np.float64(line_contents[-1][2:4])
-            second[i] = np.float64(line_contents[-1][4:])
-    # Version 1 of ATM QFIT files (binary)
-    elif (SFX == 'qi'):
-        # read input QFIT data file and subset if specified
-        fid,h = ATM1b_QFIT.read_ATM1b_QFIT_binary(input_file, SUBSETTER=s)
-        file_lines, = fid['elevation'].shape
-        ATM_L1b_input['lat'] = fid['latitude'][:]
-        ATM_L1b_input['lon'] = fid['longitude'][:]
-        ATM_L1b_input['data'] = fid['elevation'][:]
-        time_hhmmss = fid['time_hhmmss'][:]
-        # extract hour, minute and second from time_hhmmss
-        hour = np.zeros((file_lines),dtype=np.float64)
-        minute = np.zeros((file_lines),dtype=np.float64)
-        second = np.zeros((file_lines),dtype=np.float64)
-        # for each line within the file
-        for i,packed_time in enumerate(time_hhmmss):
-            # convert to zero-padded string with 3 decimal points
-            line_contents = f'{packed_time:010.3f}'
-            hour[i] = np.float64(line_contents[:2])
-            minute[i] = np.float64(line_contents[2:4])
-            second[i] = np.float64(line_contents[4:])
-    # Version 2 of ATM QFIT files (HDF5)
-    elif (SFX == 'h5'):
-        # Open the HDF5 file for reading
-        fileID = h5py.File(input_file, 'r')
-        # if not using the input subsetter
-        if s is None:
-            # create output variables with length equal to input elevation
-            file_lines, = fileID['elevation'].shape
-            ATM_L1b_input['lat'] = fileID['latitude'][:]
-            ATM_L1b_input['lon'] = fileID['longitude'][:]
-            ATM_L1b_input['data'] = fileID['elevation'][:]
-            time_hhmmss = fileID['instrument_parameters']['time_hhmmss'][:]
-        else:
-            # use subsetted data
-            file_lines = len(s)
-            ATM_L1b_input['lat'] = fileID['latitude'][s]
-            ATM_L1b_input['lon'] = fileID['longitude'][s]
-            ATM_L1b_input['data'] = fileID['elevation'][s]
-            time_hhmmss = fileID['instrument_parameters']['time_hhmmss'][s]
-        # extract hour, minute and second from time_hhmmss
-        hour = np.zeros((file_lines),dtype=np.float64)
-        minute = np.zeros((file_lines),dtype=np.float64)
-        second = np.zeros((file_lines),dtype=np.float64)
-        # for each line within the file
-        for i,packed_time in enumerate(time_hhmmss):
-            # convert to zero-padded string with 3 decimal points
-            line_contents = f'{packed_time:010.3f}'
-            hour[i] = np.float64(line_contents[:2])
-            minute[i] = np.float64(line_contents[2:4])
-            second[i] = np.float64(line_contents[4:])
-        fileID.close()
-    # leap seconds for converting from GPS time to UTC
-    S = calc_GPS_to_UTC(year,month,day,hour,minute,second)
-    # calculation of Julian day taking into account leap seconds
-    JD = calc_julian_day(year,month,day,HOUR=hour,MINUTE=minute,SECOND=second-S)
-    # converting to J2000 seconds
-    ATM_L1b_input['time'] = (JD - 2451545.0)*86400.0
-    # return the output variables
-    return ATM_L1b_input,file_lines
-
-# PURPOSE: read the ATM Level-2 data file for variables of interest
-def read_ATM_icessn_file(input_file, s):
-    # regular expression pattern for extracting parameters
-    regex_pattern = r'(BLATM2|ILATM2)_(\d+)_(\d+)_smooth_nadir(.*?)(csv|seg|pt)$'
-    # extract mission and other parameters from filename
-    MISSION,YYMMDD,HHMMSS,AUX,SFX = re.findall(regex_pattern,input_file).pop()
-    # early date strings omitted century and millenia (e.g. 93 for 1993)
-    if (len(YYMMDD) == 6):
-        yr2d,month,day = np.array([YYMMDD[:2],YYMMDD[2:4],YYMMDD[4:]],dtype='i')
-        year = (yr2d + 1900.0) if (yr2d >= 90) else (yr2d + 2000.0)
-    elif (len(YYMMDD) == 8):
-        year,month,day = np.array([YYMMDD[:4],YYMMDD[4:6],YYMMDD[6:]],dtype='i')
-    # input file column names for variables of interest with column indices
-    # variables not used: (SNslope:4, WEslope:5, npt_used:7, npt_edit:8, d:9)
-    file_dtype = {'seconds':0, 'lat':1, 'lon':2, 'data':3, 'RMS':6, 'track':-1}
-    # compile regular expression operator for reading lines (extracts numbers)
-    regex_pattern = r'[-+]?(?:(?:\d*\.\d+)|(?:\d+\.?))(?:[Ee][+-]?\d+)?'
-    rx = re.compile(regex_pattern, re.VERBOSE)
-    # read the input file, split at lines and remove all commented lines
-    with open(input_file, mode='r', encoding='utf8') as f:
-        file_contents=[i for i in f.readlines() if re.match(r'^(?!\#|\n)',i)]
-    # number of lines of data within file
-    file_lines = file_length(input_file,s)
-    # subset the data to indices if specified
-    file_contents = [file_contents[i] for i in s] if s else file_contents
-    # output python dictionary with variables
-    ATM_L2_input = {}
-    # create output variables with length equal to the number of file lines
-    for key in file_dtype.keys():
-        ATM_L2_input[key] = np.zeros((file_lines), dtype=np.float64)
-    # for each line within the file
-    for line_number,line_entries in enumerate(file_contents):
-        # find numerical instances within the line
-        line_contents = rx.findall(line_entries)
-        # for each variable of interest: save to dinput as float
-        for key,val in file_dtype.items():
-            ATM_L2_input[key][line_number] = np.float64(line_contents[val])
-    # convert shot time (seconds of day) to J2000
-    hour = np.floor(ATM_L2_input['seconds']/3600.0)
-    minute = np.floor((ATM_L2_input['seconds'] % 3600)/60.0)
-    second = ATM_L2_input['seconds'] % 60.0
-    # First column in Pre-IceBridge and ICESSN Version 1 files is GPS time
-    if (MISSION == 'BLATM2') or (SFX != 'csv'):
-        # leap seconds for converting from GPS time to UTC
-        S = calc_GPS_to_UTC(year,month,day,hour,minute,second)
-    else:
-        S = 0.0
-    # calculation of Julian day
-    JD = calc_julian_day(year,month,day,HOUR=hour,MINUTE=minute,SECOND=second-S)
-    # converting to J2000 seconds
-    ATM_L2_input['time'] = (JD - 2451545.0)*86400.0
-    # convert RMS from centimeters to meters
-    ATM_L2_input['error'] = ATM_L2_input['RMS']/100.0
-    # return the output variables
-    return ATM_L2_input,file_lines
-
-# PURPOSE: read the LVIS Level-2 data file for variables of interest
-def read_LVIS_HDF5_file(input_file, input_subsetter):
-    # LVIS region flags: GL for Greenland and AQ for Antarctica
-    lvis_flag = {'GL':'N','AQ':'S'}
-    # regular expression pattern for extracting parameters from HDF5 files
-    # computed in read_icebridge_lvis.py
-    mission_flag = r'(BLVIS2|BVLIS2|ILVIS2|ILVGH2)'
-    regex_pattern = rf'{mission_flag}_(.*?)(\d+)_(\d+)_(R\d+)_(\d+).H5'
-    # extract mission, region and other parameters from filename
-    MISSION,REGION,YY,MMDD,RLD,SS = re.findall(regex_pattern,input_file).pop()
-    LDS_VERSION = '2.0.2' if (int(RLD[1:3]) >= 18) else '1.04'
-    # input and output python dictionaries with variables
-    file_input = {}
-    LVIS_L2_input = {}
-    fileID = h5py.File(input_file,'r')
-    # create output variables with length equal to input shot number
-    file_lines = file_length(input_file,input_subsetter,HDF5='Shot_Number')
-    # https://lvis.gsfc.nasa.gov/Data/Data_Structure/DataStructure_LDS104.html
-    # https://lvis.gsfc.nasa.gov/Data/Data_Structure/DataStructure_LDS202.html
-    if (LDS_VERSION == '1.04'):
-        # elevation surfaces
-        file_input['elev'] = fileID['Elevation_Surfaces/Elevation_Centroid'][:]
-        file_input['elev_low'] = fileID['Elevation_Surfaces/Elevation_Low'][:]
-        file_input['elev_high'] = fileID['Elevation_Surfaces/Elevation_High'][:]
-        # latitude
-        file_input['lat'] = fileID['Geolocation/Latitude_Centroid'][:]
-        file_input['lat_low'] = fileID['Geolocation/Latitude_Low'][:]
-        # longitude
-        file_input['lon'] = fileID['Geolocation/Longitude_Centroid'][:]
-        file_input['lon_low'] = fileID['Geolocation/Longitude_Low'][:]
-    elif (LDS_VERSION == '2.0.2'):
-        # elevation surfaces
-        file_input['elev_low'] = fileID['Elevation_Surfaces/Elevation_Low'][:]
-        file_input['elev_high'] = fileID['Elevation_Surfaces/Elevation_High'][:]
-        # heights above lowest detected mode
-        file_input['RH50'] = fileID['Waveform/RH50'][:]
-        file_input['RH100'] = fileID['Waveform/RH100'][:]
-        # calculate centroidal elevation using 50% of waveform energy
-        file_input['elev'] = file_input['elev_low'] + file_input['RH50']
-        # latitude
-        file_input['lat_top'] = fileID['Geolocation/Latitude_Top'][:]
-        file_input['lat_low'] = fileID['Geolocation/Latitude_Low'][:]
-        # longitude
-        file_input['lon_top'] = fileID['Geolocation/Longitude_Top'][:]
-        file_input['lon_low'] = fileID['Geolocation/Longitude_Low'][:]
-        # linearly interpolate latitude and longitude to RH50
-        file_input['lat'] = file_input['lat_low'] + file_input['RH50'] * \
-            (file_input['lat_top'] - file_input['lat_low'])/file_input['RH100']
-        file_input['lon'] = file_input['lon_low'] + file_input['RH50'] * \
-            (file_input['lon_top'] - file_input['lon_low'])/file_input['RH100']
-    # J2000 seconds
-    LVIS_L2_input['time'] = fileID['Time/J2000'][:]
-    # close the input HDF5 file
-    fileID.close()
-    # output combined variables
-    LVIS_L2_input['data'] = np.zeros_like(file_input['elev'],dtype=np.float64)
-    LVIS_L2_input['lon'] = np.zeros_like(file_input['elev'],dtype=np.float64)
-    LVIS_L2_input['lat'] = np.zeros_like(file_input['elev'],dtype=np.float64)
-    LVIS_L2_input['error'] = np.zeros_like(file_input['elev'],dtype=np.float64)
-    # find where elev high is equal to elev low
-    # see note about using LVIS centroid elevation product
-    # http://lvis.gsfc.nasa.gov/OIBDataStructure.html
-    ii = np.nonzero(file_input['elev_low'] == file_input['elev_high'])
-    jj = np.nonzero(file_input['elev_low'] != file_input['elev_high'])
-    # where lowest point of waveform is equal to highest point -->
-    # using the elev_low elevation
-    LVIS_L2_input['data'][ii] = file_input['elev_low'][ii]
-    # for other locations use the centroid elevation
-    # as the centroid is a useful product over rough terrain
-    # when you are calculating ice volume change
-    LVIS_L2_input['data'][jj] = file_input['elev'][jj]
-    # latitude and longitude for each case
-    # elevation low == elevation high
-    LVIS_L2_input['lon'][ii] = file_input['lon_low'][ii]
-    LVIS_L2_input['lat'][ii] = file_input['lat_low'][ii]
-    # centroid elevations
-    LVIS_L2_input['lon'][jj] = file_input['lon'][jj]
-    LVIS_L2_input['lat'][jj] = file_input['lat'][jj]
-    # estimated uncertainty for both cases
-    LVIS_variance_low = (file_input['elev_low'] - file_input['elev'])**2
-    LVIS_variance_high = (file_input['elev_high'] - file_input['elev'])**2
-    LVIS_L2_input['error']=np.sqrt((LVIS_variance_low + LVIS_variance_high)/2.0)
-    # subset the data to indices if specified
-    if input_subsetter:
-        for key,val in LVIS_L2_input.items():
-            LVIS_L2_input[key] = val[input_subsetter]
-    # return the output variables
-    return LVIS_L2_input,file_lines,lvis_flag[REGION]
-
-# PURPOSE: calculate the Julian day from calendar date
-# http://scienceworld.wolfram.com/astronomy/JulianDate.html
-def calc_julian_day(YEAR, MONTH, DAY, HOUR=0, MINUTE=0, SECOND=0):
-    JD = 367.*YEAR - np.floor(7.*(YEAR + np.floor((MONTH+9.)/12.))/4.) - \
-        np.floor(3.*(np.floor((YEAR + (MONTH - 9.)/7.)/100.) + 1.)/4.) + \
-        np.floor(275.*MONTH/9.) + DAY + 1721028.5 + HOUR/24. + MINUTE/1440. + \
-        SECOND/86400.
-    return np.array(JD,dtype=np.float64)
-
-# PURPOSE: calculate the number of leap seconds between GPS time (seconds
-# since Jan 6, 1980 00:00:00) and UTC
-def calc_GPS_to_UTC(YEAR, MONTH, DAY, HOUR, MINUTE, SECOND):
-    GPS = 367.*YEAR - np.floor(7.*(YEAR + np.floor((MONTH+9.)/12.))/4.) - \
-        np.floor(3.*(np.floor((YEAR + (MONTH - 9.)/7.)/100.) + 1.)/4.) + \
-        np.floor(275.*MONTH/9.) + DAY + 1721028.5 - 2444244.5
-    GPS_Time = GPS*86400.0 + HOUR*3600.0 + MINUTE*60.0 + SECOND
-    return is2tk.time.count_leap_seconds(GPS_Time)
 
 # PURPOSE: convert time from delta seconds into Julian and year-decimal
 def convert_delta_time(delta_time, epoch=(2000,1,1,12,0,0)):
@@ -606,8 +312,8 @@ def main():
     REG1,TYPE1,OIB1,YMDS1,OIB2,YMD2 = rx1.findall(input_files[0]).pop()
     HEM = set_hemisphere(REG1)
     # get the dimensions of the input HDF5 file (sum auxiliary files)
-    n_1 = file_length(input_files[0], None, HDF5='data')
-    n_2 = np.sum([file_length(f,s,HDF5='data') for f,s in
+    n_1 = gz.io.icebridge.file_length(input_files[0], None, HDF5='data')
+    n_2 = np.sum([gz.io.icebridge.file_length(f,s,HDF5='data') for f,s in
         zip(input_files[1:],input_subsetter[1:])], dtype=np.int64)
 
     # full path for data directory
@@ -649,13 +355,13 @@ def main():
         # read the original data file
         if (OIB1 == 'ATM'):
             # load IceBridge ATM data
-            dinput2,file_lines = read_ATM_icessn_file(FILENAME, None)
+            dinput2,file_lines = gz.io.icebridge.read_ATM_icessn_file(FILENAME, None)
         elif (OIB1 == 'ATM1b'):
             # load IceBridge Level-1b ATM data
-            dinput2,file_lines = read_ATM_qfit_file(FILENAME, None)
+            dinput2,file_lines = gz.io.icebridge.read_ATM_qfit_file(FILENAME, None)
         elif OIB1 in ('LVIS','LVGH'):
             # load IceBridge LVIS data
-            dinput2,file_lines = read_LVIS_HDF5_file(FILENAME, None)
+            dinput2,file_lines = gz.io.icebridge.read_LVIS_HDF5_file(FILENAME, None)
         # check that data sizes match
         if (file_lines != n_1):
             logging.critical(input_files[0].name, FILENAME.name)
@@ -672,7 +378,7 @@ def main():
             c = 0
             for fi,s in zip(input_files[1:], input_subsetter[1:]):
                 file_input,att3 = read_HDF5_triangle_data(fi,s)
-                n_3 = file_length(fi,s,HDF5='data')
+                n_3 = gz.io.icebridge.file_length(fi,s,HDF5='data')
                 # iterate through input keys of iterest
                 for key in ['data','lon','lat','time','error']:
                     dinput3[key][c:c+n_3] = file_input.get(key,None)
@@ -686,13 +392,13 @@ def main():
                 # read the original data file
                 if (OIB1 == 'ATM'):
                     # load IceBridge ATM data from FILENAME
-                    file_input,n_4 = read_ATM_icessn_file(FILENAME,s)
+                    file_input,n_4 = gz.io.icebridge.read_ATM_icessn_file(FILENAME,s)
                 elif (OIB1 == 'ATM1b'):
                     # load IceBridge Level-1b ATM data from FILENAME
-                    file_input,n_4 = read_ATM_qfit_file(FILENAME,s)
+                    file_input,n_4 = gz.io.icebridge.read_ATM_qfit_file(FILENAME,s)
                 elif OIB1 in ('LVIS','LVGH'):
                     # load IceBridge LVIS data from FILENAME
-                    file_input,n_4 = read_LVIS_HDF5_file(FILENAME,s)
+                    file_input,n_4 = gz.io.icebridge.read_LVIS_HDF5_file(FILENAME,s)
                 # check that data sizes match
                 if (n_3 != n_4):
                     logging.critical(fi.name, FILENAME.name)
