+++ tides/compute_LPT_icebridge_data.py
@@ -39,6 +39,7 @@ UPDATE HISTORY:
     Updated 05/2023: use timescale class for time conversion operations
         use defaults from eop module for pole tide and EOP files
         using pathlib to define and operate on paths
+        move icebridge data inputs to a separate module in io
     Updated 03/2023: added option for changing the IERS mean pole convention
     Updated 12/2022: single implicit import of grounding zone tools
         use constants class from pyTMD for ellipsoidal parameters
@@ -79,11 +80,6 @@ import grounding_zones as gz
 
 # attempt imports
 try:
-    import ATM1b_QFIT.read_ATM1b_QFIT_binary
-except (ImportError, ModuleNotFoundError) as exc:
-    warnings.filterwarnings("module")
-    warnings.warn("ATM1b_QFIT not available", ImportWarning)
-try:
     import h5py
 except (ImportError, ModuleNotFoundError) as exc:
     warnings.filterwarnings("module")
@@ -95,312 +91,6 @@ except (ImportError, ModuleNotFoundError) as exc:
     warnings.warn("pyTMD not available", ImportWarning)
 # ignore warnings
 warnings.filterwarnings("ignore")
-
-# PURPOSE: reading the number of file lines removing commented lines
-def file_length(input_file, input_subsetter, HDF5=False, QFIT=False):
-    # verify input file is path
-    input_file = pathlib.Path(input_file).expanduser().absolute()
-    # subset the data to indices if specified
-    if input_subsetter:
-        file_lines = len(input_subsetter)
-    elif HDF5:
-        # read the size of an input variable within a HDF5 file
-        with h5py.File(input_file, 'r') as fileID:
-            file_lines, = fileID[HDF5].shape
-    elif QFIT:
-        # read the size of a QFIT binary file
-        file_lines = ATM1b_QFIT.ATM1b_QFIT_shape(input_file)
-    else:
-        # read the input file, split at lines and remove all commented lines
-        with input_file.open(mode='r', encoding='utf8') as f:
-            i = [i for i in f.readlines() if re.match(r'^(?!\#|\n)',i)]
-        file_lines = len(i)
-    # return the number of lines
-    return file_lines
-
-## PURPOSE: read the ATM Level-1b data file for variables of interest
-def read_ATM_qfit_file(input_file, input_subsetter):
-    # verify input file is path
-    input_file = pathlib.Path(input_file).expanduser().absolute()
-    # regular expression pattern for extracting parameters
-    mission_flag = r'(BLATM1B|ILATM1B|ILNSA1B)'
-    regex_pattern = rf'{mission_flag}_(\d+)_(\d+)(.*?).(qi|TXT|h5)'
-    regex = re.compile(regex_pattern, re.VERBOSE)
-    # extract mission and other parameters from filename
-    MISSION,YYMMDD,HHMMSS,AUX,SFX = regex.findall(input_file.name).pop()
-    # early date strings omitted century and millenia (e.g. 93 for 1993)
-    if (len(YYMMDD) == 6):
-        yr2d,month,day = np.array([YYMMDD[:2],YYMMDD[2:4],YYMMDD[4:]],dtype='i')
-        year = (yr2d + 1900.0) if (yr2d >= 90) else (yr2d + 2000.0)
-    elif (len(YYMMDD) == 8):
-        year,month,day = np.array([YYMMDD[:4],YYMMDD[4:6],YYMMDD[6:]],dtype='i')
-    # output python dictionary with variables
-    ATM_L1b_input = {}
-    # Version 1 of ATM QFIT files (ascii)
-    # output text file from qi2txt with proper filename format
-    # do not use the shortened output format from qi2txt
-    if (SFX == 'TXT'):
-        # compile regular expression operator for reading lines
-        regex_pattern = r'[-+]?(?:(?:\d*\.\d+)|(?:\d+\.?))(?:[Ee][+-]?\d+)?'
-        rx = re.compile(regex_pattern, re.VERBOSE)
-        # read the input file, split at lines and remove all commented lines
-        with input_file.open(mode='r', encoding='utf8') as f:
-            file_contents = [i for i in f.read().splitlines() if
-                re.match(r'^(?!\#|\n)',i)]
-        # number of lines of data within file
-        file_lines = file_length(input_file, input_subsetter)
-        # create output variables with length equal to the number of lines
-        ATM_L1b_input['lat'] = np.zeros_like(file_contents, dtype=np.float64)
-        ATM_L1b_input['lon'] = np.zeros_like(file_contents, dtype=np.float64)
-        ATM_L1b_input['data'] = np.zeros_like(file_contents, dtype=np.float64)
-        hour = np.zeros_like(file_contents, dtype=np.float64)
-        minute = np.zeros_like(file_contents, dtype=np.float64)
-        second = np.zeros_like(file_contents, dtype=np.float64)
-        # for each line within the file
-        for i,line in enumerate(file_contents):
-            # find numerical instances within the line
-            line_contents = rx.findall(line)
-            ATM_L1b_input['lat'][i] = np.float64(line_contents[1])
-            ATM_L1b_input['lon'][i] = np.float64(line_contents[2])
-            ATM_L1b_input['data'][i] = np.float64(line_contents[3])
-            hour[i] = np.float64(line_contents[-1][:2])
-            minute[i] = np.float64(line_contents[-1][2:4])
-            second[i] = np.float64(line_contents[-1][4:])
-    # Version 1 of ATM QFIT files (binary)
-    elif (SFX == 'qi'):
-        # read input QFIT data file and subset if specified
-        fid, h = ATM1b_QFIT.read_ATM1b_QFIT_binary(input_file)
-        # number of lines of data within file
-        file_lines = file_length(input_file, input_subsetter, QFIT=True)
-        ATM_L1b_input['lat'] = fid['latitude'][:]
-        ATM_L1b_input['lon'] = fid['longitude'][:]
-        ATM_L1b_input['data'] = fid['elevation'][:]
-        time_hhmmss = fid['time_hhmmss'][:]
-        # extract hour, minute and second from time_hhmmss
-        hour = np.zeros_like(time_hhmmss, dtype=np.float64)
-        minute = np.zeros_like(time_hhmmss, dtype=np.float64)
-        second = np.zeros_like(time_hhmmss, dtype=np.float64)
-        # for each line within the file
-        for i,packed_time in enumerate(time_hhmmss):
-            # convert to zero-padded string with 3 decimal points
-            line_contents = f'{packed_time:010.3f}'
-            hour[i] = np.float64(line_contents[:2])
-            minute[i] = np.float64(line_contents[2:4])
-            second[i] = np.float64(line_contents[4:])
-    # Version 2 of ATM QFIT files (HDF5)
-    elif (SFX == 'h5'):
-        # Open the HDF5 file for reading
-        fileID = h5py.File(input_file, 'r')
-        # number of lines of data within file
-        file_lines = file_length(input_file, input_subsetter,
-            HDF5='elevation')
-        # create output variables with length equal to input elevation
-        ATM_L1b_input['lat'] = fileID['latitude'][:]
-        ATM_L1b_input['lon'] = fileID['longitude'][:]
-        ATM_L1b_input['data'] = fileID['elevation'][:]
-        time_hhmmss = fileID['instrument_parameters']['time_hhmmss'][:]
-        # extract hour, minute and second from time_hhmmss
-        hour = np.zeros_like(time_hhmmss, dtype=np.float64)
-        minute = np.zeros_like(time_hhmmss, dtype=np.float64)
-        second = np.zeros_like(time_hhmmss, dtype=np.float64)
-        # for each line within the file
-        for i,packed_time in enumerate(time_hhmmss):
-            # convert to zero-padded string with 3 decimal points
-            line_contents = f'{packed_time:010.3f}'
-            hour[i] = np.float64(line_contents[:2])
-            minute[i] = np.float64(line_contents[2:4])
-            second[i] = np.float64(line_contents[4:])
-        # close the input HDF5 file
-        fileID.close()
-    # calculate the number of leap seconds between GPS time (seconds
-    # since Jan 6, 1980 00:00:00) and UTC
-    gps_seconds = pyTMD.time.convert_calendar_dates(year,month,day,
-        hour=hour,minute=minute,second=second,
-        epoch=pyTMD.time._gps_epoch,scale=86400.0)
-    leap_seconds = pyTMD.time.count_leap_seconds(gps_seconds)
-    # calculation of Julian day taking into account leap seconds
-    # converting to J2000 seconds
-    ATM_L1b_input['time'] = pyTMD.time.convert_calendar_dates(year,month,day,
-        hour=hour,minute=minute,second=second-leap_seconds,
-        epoch=pyTMD.time._j2000_epoch,scale=86400.0)
-    # subset the data to indices if specified
-    if input_subsetter:
-        for key,val in ATM_L1b_input.items():
-            ATM_L1b_input[key] = val[input_subsetter]
-    # hemispheric shot count
-    count = {}
-    count['N'] = np.count_nonzero(ATM_L1b_input['lat'] >= 0.0)
-    count['S'] = np.count_nonzero(ATM_L1b_input['lat'] < 0.0)
-    # determine hemisphere with containing shots in file
-    HEM, = [key for key, val in count.items() if val]
-    # return the output variables
-    return ATM_L1b_input, file_lines, HEM
-
-# PURPOSE: read the ATM Level-2 data file for variables of interest
-def read_ATM_icessn_file(input_file, input_subsetter):
-    # verify input file is path
-    input_file = pathlib.Path(input_file).expanduser().absolute()
-    # regular expression pattern for extracting parameters
-    mission_flag = r'(BLATM2|ILATM2)'
-    regex_pattern = rf'{mission_flag}_(\d+)_(\d+)_smooth_nadir(.*?)(csv|seg|pt)$'
-    regex = re.compile(regex_pattern, re.VERBOSE)
-    # extract mission and other parameters from filename
-    MISSION,YYMMDD,HHMMSS,AUX,SFX = regex.findall(input_file.name).pop()
-    # early date strings omitted century and millenia (e.g. 93 for 1993)
-    if (len(YYMMDD) == 6):
-        yr2d,month,day = np.array([YYMMDD[:2],YYMMDD[2:4],YYMMDD[4:]],dtype='i')
-        year = (yr2d + 1900.0) if (yr2d >= 90) else (yr2d + 2000.0)
-    elif (len(YYMMDD) == 8):
-        year,month,day = np.array([YYMMDD[:4],YYMMDD[4:6],YYMMDD[6:]],dtype='i')
-    # input file column names for variables of interest with column indices
-    # variables not used: (SNslope:4, WEslope:5, npt_used:7, npt_edit:8, d:9)
-    file_dtype = {'seconds':0, 'lat':1, 'lon':2, 'data':3, 'RMS':6, 'track':-1}
-    # compile regular expression operator for reading lines (extracts numbers)
-    regex_pattern = r'[-+]?(?:(?:\d*\.\d+)|(?:\d+\.?))(?:[Ee][+-]?\d+)?'
-    rx = re.compile(regex_pattern, re.VERBOSE)
-    # read the input file, split at lines and remove all commented lines
-    with open(input_file, mode='r', encoding='utf8') as f:
-        file_contents = [i for i in f.read().splitlines() if
-            re.match(r'^(?!\#|\n)',i)]
-    # number of lines of data within file
-    file_lines = file_length(input_file, input_subsetter)
-    # output python dictionary with variables
-    ATM_L2_input = {}
-    # create output variables with length equal to the number of file lines
-    for key in file_dtype.keys():
-        ATM_L2_input[key] = np.zeros_like(file_contents, dtype=np.float64)
-    # for each line within the file
-    for line_number,line_entries in enumerate(file_contents):
-        # find numerical instances within the line
-        line_contents = rx.findall(line_entries)
-        # for each variable of interest: save to dinput as float
-        for key,val in file_dtype.items():
-            ATM_L2_input[key][line_number] = np.float64(line_contents[val])
-    # convert shot time (seconds of day) to J2000
-    hour = np.floor(ATM_L2_input['seconds']/3600.0)
-    minute = np.floor((ATM_L2_input['seconds'] % 3600)/60.0)
-    second = ATM_L2_input['seconds'] % 60.0
-    # First column in Pre-IceBridge and ICESSN Version 1 files is GPS time
-    if (MISSION == 'BLATM2') or (SFX != 'csv'):
-        # calculate the number of leap seconds between GPS time (seconds
-        # since Jan 6, 1980 00:00:00) and UTC
-        gps_seconds = pyTMD.time.convert_calendar_dates(year,month,day,
-            hour=hour,minute=minute,second=second,
-            epoch=pyTMD.time._gps_epoch,scale=86400.0)
-        leap_seconds = pyTMD.time.count_leap_seconds(gps_seconds)
-    else:
-        leap_seconds = 0.0
-    # calculation of Julian day
-    # converting to J2000 seconds
-    ATM_L2_input['time'] = pyTMD.time.convert_calendar_dates(year,month,day,
-        hour=hour,minute=minute,second=second-leap_seconds,
-        epoch=pyTMD.time._j2000_epoch,scale=86400.0)
-    # convert RMS from centimeters to meters
-    ATM_L2_input['error'] = ATM_L2_input['RMS']/100.0
-    # subset the data to indices if specified
-    if input_subsetter:
-        for key,val in ATM_L2_input.items():
-            ATM_L2_input[key] = val[input_subsetter]
-    # hemispheric shot count
-    count = {}
-    count['N'] = np.count_nonzero(ATM_L2_input['lat'] >= 0.0)
-    count['S'] = np.count_nonzero(ATM_L2_input['lat'] < 0.0)
-    # determine hemisphere with containing shots in file
-    HEM, = [key for key, val in count.items() if val]
-    # return the output variables
-    return ATM_L2_input, file_lines, HEM
-
-# PURPOSE: read the LVIS Level-2 data file for variables of interest
-def read_LVIS_HDF5_file(input_file, input_subsetter):
-    # verify input file is path
-    input_file = pathlib.Path(input_file).expanduser().absolute()
-    # LVIS region flags: GL for Greenland and AQ for Antarctica
-    lvis_flag = {'GL':'N','AQ':'S'}
-    # regular expression pattern for extracting parameters from HDF5 files
-    # computed in read_icebridge_lvis.py
-    mission_flag = r'(BLVIS2|BVLIS2|ILVIS2|ILVGH2)'
-    regex_pattern = rf'{mission_flag}_(.*?)(\d+)_(\d+)_(R\d+)_(\d+).H5'
-    regex = re.compile(regex_pattern, re.VERBOSE)
-    # extract mission, region and other parameters from filename
-    MISSION,REGION,YY,MMDD,RLD,SS = regex.findall(input_file.name).pop()
-    LDS_VERSION = '2.0.2' if (int(RLD[1:3]) >= 18) else '1.04'
-    # input and output python dictionaries with variables
-    file_input = {}
-    LVIS_L2_input = {}
-    fileID = h5py.File(input_file, 'r')
-    # create output variables with length equal to input shot number
-    file_lines = file_length(input_file, input_subsetter, HDF5='Shot_Number')
-    # https://lvis.gsfc.nasa.gov/Data/Data_Structure/DataStructure_LDS104.html
-    # https://lvis.gsfc.nasa.gov/Data/Data_Structure/DataStructure_LDS202.html
-    if (LDS_VERSION == '1.04'):
-        # elevation surfaces
-        file_input['elev'] = fileID['Elevation_Surfaces/Elevation_Centroid'][:]
-        file_input['elev_low'] = fileID['Elevation_Surfaces/Elevation_Low'][:]
-        file_input['elev_high'] = fileID['Elevation_Surfaces/Elevation_High'][:]
-        # latitude
-        file_input['lat'] = fileID['Geolocation/Latitude_Centroid'][:]
-        file_input['lat_low'] = fileID['Geolocation/Latitude_Low'][:]
-        # longitude
-        file_input['lon'] = fileID['Geolocation/Longitude_Centroid'][:]
-        file_input['lon_low'] = fileID['Geolocation/Longitude_Low'][:]
-    elif (LDS_VERSION == '2.0.2'):
-        # elevation surfaces
-        file_input['elev_low'] = fileID['Elevation_Surfaces/Elevation_Low'][:]
-        file_input['elev_high'] = fileID['Elevation_Surfaces/Elevation_High'][:]
-        # heights above lowest detected mode
-        file_input['RH50'] = fileID['Waveform/RH50'][:]
-        file_input['RH100'] = fileID['Waveform/RH100'][:]
-        # calculate centroidal elevation using 50% of waveform energy
-        file_input['elev'] = file_input['elev_low'] + file_input['RH50']
-        # latitude
-        file_input['lat_top'] = fileID['Geolocation/Latitude_Top'][:]
-        file_input['lat_low'] = fileID['Geolocation/Latitude_Low'][:]
-        # longitude
-        file_input['lon_top'] = fileID['Geolocation/Longitude_Top'][:]
-        file_input['lon_low'] = fileID['Geolocation/Longitude_Low'][:]
-        # linearly interpolate latitude and longitude to RH50
-        file_input['lat'] = file_input['lat_low'] + file_input['RH50'] * \
-            (file_input['lat_top'] - file_input['lat_low'])/file_input['RH100']
-        file_input['lon'] = file_input['lon_low'] + file_input['RH50'] * \
-            (file_input['lon_top'] - file_input['lon_low'])/file_input['RH100']
-    # J2000 seconds
-    LVIS_L2_input['time'] = fileID['Time/J2000'][:]
-    # close the input HDF5 file
-    fileID.close()
-    # output combined variables
-    LVIS_L2_input['data'] = np.zeros_like(file_input['elev'], dtype=np.float64)
-    LVIS_L2_input['lon'] = np.zeros_like(file_input['elev'], dtype=np.float64)
-    LVIS_L2_input['lat'] = np.zeros_like(file_input['elev'], dtype=np.float64)
-    LVIS_L2_input['error'] = np.zeros_like(file_input['elev'], dtype=np.float64)
-    # find where elev high is equal to elev low
-    # see note about using LVIS centroid elevation product
-    # http://lvis.gsfc.nasa.gov/OIBDataStructure.html
-    ii = np.nonzero(file_input['elev_low'] == file_input['elev_high'])
-    jj = np.nonzero(file_input['elev_low'] != file_input['elev_high'])
-    # where lowest point of waveform is equal to highest point -->
-    # using the elev_low elevation
-    LVIS_L2_input['data'][ii] = file_input['elev_low'][ii]
-    # for other locations use the centroid elevation
-    # as the centroid is a useful product over rough terrain
-    # when you are calculating ice volume change
-    LVIS_L2_input['data'][jj] = file_input['elev'][jj]
-    # latitude and longitude for each case
-    # elevation low == elevation high
-    LVIS_L2_input['lon'][ii] = file_input['lon_low'][ii]
-    LVIS_L2_input['lat'][ii] = file_input['lat_low'][ii]
-    # centroid elevations
-    LVIS_L2_input['lon'][jj] = file_input['lon'][jj]
-    LVIS_L2_input['lat'][jj] = file_input['lat'][jj]
-    # estimated uncertainty for both cases
-    LVIS_variance_low = (file_input['elev_low'] - file_input['elev'])**2
-    LVIS_variance_high = (file_input['elev_high'] - file_input['elev'])**2
-    LVIS_L2_input['error']=np.sqrt((LVIS_variance_low + LVIS_variance_high)/2.0)
-    # subset the data to indices if specified
-    if input_subsetter:
-        for key,val in LVIS_L2_input.items():
-            LVIS_L2_input[key] = val[input_subsetter]
-    # return the output variables
-    return LVIS_L2_input, file_lines, lvis_flag[REGION]
 
 # PURPOSE: read Operation IceBridge data from NSIDC
 # compute load pole tide radial displacements at data points and times
@@ -490,13 +180,16 @@ def compute_LPT_icebridge_data(arg, CONVENTION='2018', VERBOSE=False, MODE=0o775
     logger.info(f'{str(input_file)} -->')
     if (OIB == 'ATM'):
         # load IceBridge ATM data from input_file
-        dinput,file_lines,HEM = read_ATM_icessn_file(input_file, input_subsetter)
+        dinput, file_lines, HEM = gz.io.icebridge.read_ATM_icessn_file(
+            input_file, input_subsetter)
     elif (OIB == 'ATM1b'):
         # load IceBridge Level-1b ATM data from input_file
-        dinput,file_lines,HEM = read_ATM_qfit_file(input_file, input_subsetter)
+        dinput, file_lines, HEM = gz.io.icebridge.read_ATM_qfit_file(
+            input_file, input_subsetter)
     elif OIB in ('LVIS','LVGH'):
         # load IceBridge LVIS data from input_file
-        dinput,file_lines,HEM = read_LVIS_HDF5_file(input_file, input_subsetter)
+        dinput, file_lines, HEM = gz.io.icebridge.read_LVIS_HDF5_file(
+            input_file, input_subsetter)
 
     # extract lat/lon
     lon = dinput['lon'][:]
